refactor test code

+++ src/datum/datum.rs
@@ -14,7 +14,7 @@ pub trait DatumTrait: Display + Debug + Clone {
     fn from_raw(raw: &str) -> Self {
         let l = raw.len();
         println!("raw: {}, l: {}", raw, l);
-        if raw.chars().nth(0 as usize).unwrap() == '"'
+        if raw.chars().nth(0_usize).unwrap() == '"'
             && raw.chars().nth((l - 1) as usize).unwrap() == '"'
         {
             let inner = &raw[1..l - 1];
